requirements.txt
pyvirtualdisplay==1.3.2
selenium==3.141.0
python-dotenv==0.14.0
Flask==1.1.2
shadow-useragent==0.0.17
pyscreenshot==2.2
Pillow==7.2.0
xvfbwrapper==0.2.9